% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_DDM_extras.R
\name{ADDMnoC}
\alias{ADDMnoC}
\title{Advantage DDM, allowing for sv increasing with stimulus sum}
\usage{
ADDMnoC()
}
\value{
A list defining the cognitive model
}
\description{
Advantage DDM, allowing for sv increasing with stimulus sum
}
