% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LBA_extras.R
\name{ALBAnoC}
\alias{ALBAnoC}
\title{Advantage LBA Model with no C, pbeta mapping}
\usage{
ALBAnoC()
}
\value{
A list defining the cognitive model
}
\description{
Advantage LBA Model with no C, pbeta mapping
}
