% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_RDM_extras.R
\name{ARDM}
\alias{ARDM}
\title{Advantage Racing Diffusion Model, pbeta mapping}
\usage{
ARDM()
}
\value{
A list defining the cognitive model
}
\description{
Advantage Racing Diffusion Model, pbeta mapping
}
