% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_RDM_extras.R
\name{ARDMnoC}
\alias{ARDMnoC}
\title{Advantage Racing Diffusion Model with no C, pbeta mapping}
\usage{
ARDMnoC()
}
\value{
A list defining the cognitive model
}
\description{
Advantage Racing Diffusion Model with no C, pbeta mapping
}
