% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LNR_extras.R
\name{lnrMP}
\alias{lnrMP}
\title{lnr model where s proportional to m}
\usage{
lnrMP()
}
\value{
A model list with all the necessary functions to sample
}
\description{
lnr model where s proportional to m
}
