% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_DDM_extras.R
\name{MDDM}
\alias{MDDM}
\title{MDDM: Missing value DDM with contamination}
\usage{
MDDM()
}
\value{
A model list with all the necessary functions to sample
}
\description{
MDDM: Missing value DDM with contamination
}
