% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LBA_extras.R
\name{MLBA}
\alias{MLBA}
\title{MLBA: Missing parameterization for LBA}
\usage{
MLBA()
}
\value{
A model list with all the necessary functions to sample
}
\description{
LBA model accommodating missing values (truncation and censoring) and
assuming positive rates (i.e., no intrinsic omissions)
}
