% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_RDM_extras.R
\name{MRDM}
\alias{MRDM}
\title{MRDM: RDM parameterization with missing values}
\usage{
MRDM()
}
\value{
A list defining the cognitive model
}
\description{
MRDM: RDM parameterization with missing values
}
