% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRI.R
\name{normal_mri}
\alias{normal_mri}
\title{Run an MRI model}
\usage{
normal_mri()
}
\value{
A model list with all the relevant functions to run the MRI model
}
\description{
Run an MRI model
}
\examples{
normal_mri()
}
