% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_srrt}
\alias{plot_srrt}
\title{Plot signal-response response times (SRRT)}
\usage{
plot_srrt(data, breaks = 5, subject = NULL, pp = NULL)
}
\arguments{
\item{data}{A data frame. Needs to have the variables \code{subjects}, \code{rt}, \code{R}, and \code{SSD}.}

\item{breaks}{Number of breaks for the stop-signal delay (SSD).}

\item{subject}{If set to \code{NULL} (default), averaged data is plotted, otherwise individual plots are created.}

\item{pp}{A data frame with posterior predictives as returned by predict.emc()}
}
\value{
A matrix with the plotted data
}
\description{
Plot signal-response response times (SRRT)
}
