% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_RDM_extras.R
\name{RDMnoC}
\alias{RDMnoC}
\title{The Racing Diffusion Model (RDM) with no C}
\usage{
RDMnoC()
}
\value{
A list defining the cognitive model
}
\description{
The Racing Diffusion Model (RDM) with no C
}
