% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LBA_extras.R
\name{rLBA_BE}
\alias{rLBA_BE}
\title{#' Advantage LBA Model with no C, 3 parameter shape mapping
#'
#' f <- function(x, a, b, p) {(pbeta(x, a, b)\emph{p + qbeta(x, a, b)}(1 - p))}
#' a = S1, b = S1 + SD, both log scale, p = SW, 0<p<1 probit scale
#'
#' @return A list defining the cognitive model
#' @export
AiLBAnoC <- function(){
list(
type="RACE",
# p_vector transform, sets sv as a scaling parameter
p_types=c("v" = 0,"sv" = log(1),"B" = log(1),"A" = log(0),"t0" = log(0),
"AD"=log(1),"AS"=log(0),S1=log(1),SD=log(1),SW=qnorm(.5)),
transform = function(p) p,
# Transform to natural scale
Ntransform=function(x,use=NULL) {
if (is.null(use)) {
use <- !(names(x) \%in\% c("v","S1","SD","SW"))
x\link{,use} <- exp(x\link{,use})
} else if (!all(use=="v")) {
ok <- use\link{!(use \%in\% c("v","S1","SD","SW"))}
x\link{,ok} <- exp(x\link{,ok})
}
x
},
# Trial dependent parameter transform
Ttransform = function(pars,dadm) {
pars <- cbind(pars,b=pars\link{,"B"} + pars\link{,"A"})
if (!is.null(attr(dadm,"adaptive"))) pars <- do_adaptive(pars,dadm)
pars <- advantage_pars(pars,dadm[attr(dadm,"advantage")[\link{1}]],
length(levels(dadm$lR)),names(attr(dadm,"advantage"))\link{1},inv=TRUE)
attr(pars,"ok") <- (pars\link{,"sv"} > 0) & (pars\link{,"B"} >= 0) &
(pars\link{,"t0"} > .05) & ((pars\link{,"A"} > 1e-6) | pars\link{,"A"} == 0)
pars
},
# Random function for racing accumulator
rfun=function(lR=NULL,pars) {
ok <- (pars\link{,"sv"} > 0) & (pars\link{,"B"} >= 0) &
(pars\link{,"t0"} > .05) & ((pars\link{,"A"} > 1e-6) | pars\link{,"A"} == 0)
if (is.null(lR)) ok else rALBA(lR,pars,posdrift=TRUE,ok=ok)
},
# Density function (PDF) for single accumulator
dfun=function(rt,pars) dLBA(rt,pars,posdrift = TRUE, robust = FALSE),
# Probability function (CDF) for single accumulator
pfun=function(rt,pars) pLBA(rt,pars,posdrift = TRUE, robust = FALSE),
# Race likelihood combining pfun and dfun
log_likelihood=function(p_vector,dadm,min_ll=log(1e-10)){
log_likelihood_race_advantage(p_vector=p_vector, dadm = dadm, min_ll = min_ll)
}
)
}}
\usage{
rLBA_BE(pars, ok = rep(TRUE, dim(pars)[1]), return_activation = FALSE)
}
\description{
#' Advantage LBA Model with no C, 3 parameter shape mapping
#'
#' f <- function(x, a, b, p) {(pbeta(x, a, b)\emph{p + qbeta(x, a, b)}(1 - p))}
#' a = S1, b = S1 + SD, both log scale, p = SW, 0<p<1 probit scale
#'
#' @return A list defining the cognitive model
#' @export
AiLBAnoC <- function(){
list(
type="RACE",
# p_vector transform, sets sv as a scaling parameter
p_types=c("v" = 0,"sv" = log(1),"B" = log(1),"A" = log(0),"t0" = log(0),
"AD"=log(1),"AS"=log(0),S1=log(1),SD=log(1),SW=qnorm(.5)),
transform = function(p) p,
# Transform to natural scale
Ntransform=function(x,use=NULL) {
if (is.null(use)) {
use <- !(names(x) \%in\% c("v","S1","SD","SW"))
x\link{,use} <- exp(x\link{,use})
} else if (!all(use=="v")) {
ok <- use\link{!(use \%in\% c("v","S1","SD","SW"))}
x\link{,ok} <- exp(x\link{,ok})
}
x
},
# Trial dependent parameter transform
Ttransform = function(pars,dadm) {
pars <- cbind(pars,b=pars\link{,"B"} + pars\link{,"A"})
if (!is.null(attr(dadm,"adaptive"))) pars <- do_adaptive(pars,dadm)
pars <- advantage_pars(pars,dadm[attr(dadm,"advantage")[\link{1}]],
length(levels(dadm$lR)),names(attr(dadm,"advantage"))\link{1},inv=TRUE)
attr(pars,"ok") <- (pars\link{,"sv"} > 0) & (pars\link{,"B"} >= 0) &
(pars\link{,"t0"} > .05) & ((pars\link{,"A"} > 1e-6) | pars\link{,"A"} == 0)
pars
},
# Random function for racing accumulator
rfun=function(lR=NULL,pars) {
ok <- (pars\link{,"sv"} > 0) & (pars\link{,"B"} >= 0) &
(pars\link{,"t0"} > .05) & ((pars\link{,"A"} > 1e-6) | pars\link{,"A"} == 0)
if (is.null(lR)) ok else rALBA(lR,pars,posdrift=TRUE,ok=ok)
},
# Density function (PDF) for single accumulator
dfun=function(rt,pars) dLBA(rt,pars,posdrift = TRUE, robust = FALSE),
# Probability function (CDF) for single accumulator
pfun=function(rt,pars) pLBA(rt,pars,posdrift = TRUE, robust = FALSE),
# Race likelihood combining pfun and dfun
log_likelihood=function(p_vector,dadm,min_ll=log(1e-10)){
log_likelihood_race_advantage(p_vector=p_vector, dadm = dadm, min_ll = min_ll)
}
)
}
}
