% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_SS_EXG.R
\name{SSexGuv}
\alias{SSexGuv}
\title{Stop-signal exGaussian race with uni-valent stop-triggered responding}
\usage{
SSexGuv()
}
\value{
A model list with all the necessary functions to sample
}
\description{
Stop-signal exGaussian race with uni-valent stop-triggered responding
}
