% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_SS_EXG.R
\name{staircase_fun}
\alias{staircase_fun}
\title{Staircase algorithm}
\usage{
staircase_fun(
  dadm,
  p = 0.25,
  pars = NULL,
  SSD0 = 0.25,
  stairstep = 0.05,
  stairmin = 0,
  stairmax = Inf
)
}
\value{
A function to simulate staircase SSD data for
}
\description{
Staircase algorithm
}
