% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRI.R
\name{make_design_fmri}
\alias{make_design_fmri}
\title{Title}
\usage{
make_design_fmri(
  data,
  events,
  model = normal_mri,
  factors,
  contrasts,
  hrf_model = "glover + derivative",
  ...
)
}
\arguments{
\item{data}{The voxel or ROI timeseries (one at a time)}

\item{events}{The event file}

\item{model}{The fMRI model used. The default and only option at the moment is \code{normal_mri}}

\item{factors}{A list. The factors used in the events files, e.g. list(E = c('acc', 'spd'))}

\item{contrasts}{A list with one entry per contrast}

\item{hrf_model}{The type of convolution used. Options, \code{glover} or \code{glover + derivative}}

\item{...}{}
}
\value{
An emc design matrix that can be used in \code{make_emc}
}
\description{
Title
}
